foodCrawlingDemo

+++ foodCrawlingDemo/src/main/java/com/javateam/foodCrawlingDemo/repository/CURepository.java
@@ -1,9 +1,11 @@
 package com.javateam.foodCrawlingDemo.repository;
 
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.PagingAndSortingRepository;
+
 import com.javateam.foodCrawlingDemo.domain.CUVO;
 
-public interface CURepository extends CrudRepository<CUVO, Integer> {
+public interface CURepository extends CrudRepository<CUVO, Long> {
 	
 	
 }
